Add NoSnapshotsDriver
