+++ java/system-test/system-test-dsl/src/main/java/sleeper/systemtest/dsl/util/SystemTestDriversBase.java
@@ -27,7 +27,6 @@ import sleeper.systemtest.dsl.ingest.IngestByAnyQueueDriver;
 import sleeper.systemtest.dsl.ingest.IngestByQueue;
 import sleeper.systemtest.dsl.ingest.IngestLocalFileByAnyQueueDriver;
 import sleeper.systemtest.dsl.ingest.InvokeIngestTasksDriver;
-import sleeper.systemtest.dsl.instance.ScheduleRulesDriver;
 import sleeper.systemtest.dsl.metrics.TableMetricsDriver;
 import sleeper.systemtest.dsl.partitioning.PartitionSplittingDriver;
 import sleeper.systemtest.dsl.python.PythonQueryTypesDriver;
@@ -171,10 +170,5 @@ public abstract class SystemTestDriversBase implements SystemTestDrivers {
     @Override
     public TableMetricsDriver tableMetrics(SystemTestContext context) {
         throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public ScheduleRulesDriver schedules() {
-        return new NoScheduleRulesDriver();
     }
 }
