--- java/system-test/system-test-dsl/src/test/java/sleeper/systemtest/dsl/testutil/drivers/InMemorySnapshotsDriver.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2022-2024 Crown Copyright
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package sleeper.systemtest.dsl.testutil.drivers;
-
-import sleeper.core.properties.instance.InstanceProperties;
-import sleeper.systemtest.dsl.snapshot.SnapshotsDriver;
-
-public class InMemorySnapshotsDriver implements SnapshotsDriver {
-    public InMemorySnapshotsDriver() {
-    }
-
-    @Override
-    public void enableCreation(InstanceProperties instanceProperties) {
-    }
-
-    @Override
-    public void disableCreation(InstanceProperties instanceProperties) {
-    }
-}
